BUG: Unixhome-ingo counts does not aggregate numbers (cp. 7672dbcd)
That wrapper uses git-wcs-in-dir-do-command, but the counts extension requires an accumulationCommand.
Like git-wc-with-suffix-do, git-wcs-in-dir-do needs to use executeCommand() (which considers accumulationCommand) instead of invoking git-wcdo-core directly.
For that, we need to redefine GIT_DOEXTENSIONS_WRAPPEE and put wcs into dashdashArgs.

(This relates to commit 7672dbcde96a4e18527268618324c32b70544eb8)