Robustness: Also parse custom args before the dashdash part
My Unixhome-ingo et al. don't do this, as they provide the DIRSPEC ... and leave the actual SIMPLECOMMAND (and any custom flags like --quiet) to be passed by the user.
But the usage help actually advertises the placement at the beginning.
To handle both positions, rename parseDashdashAndUnknownArgs() into parseCustomInto() and inject the variable that will hold the non-custom args: For dashdashArgs, that'll gather the DIRSPEC(s) in dirspecs, for unknownOptions, it'll complain.