git-wrapper: ENH: Apply the alias transformation of "X" to "-x" here as well
Aliases may also invoked through git-wrapper. So far, the br... aliases directly passed on the already translated subcommand:
	tdlg-i = revision-command --keep-position today-command --revision REVISION -2 lg-i TIMESPAN
But now I receive an untranslated "git-td lgI", and would have to translate inside git-td (from lgI to lg-i), which is cumbersome.
Rather delegate this to the git-wrapper that's eventually invoked by the today-command, revision-command, ...