Generalization: Extract git-in
The incoming commits are only on a remote branch, not the local one (that's the essence of "incoming"). Therefore, any commands that use rebase or modify the current branch (removing commits) do not make sense. We handle this via a new scopeNoLocalBranch flag.

For the special implementation of "inbase", we can use the wrapper's sub-sub-command override and provide git-in-base.

FIX: in{base,pred}{lg,s} never worked; need to handle --no-range in git-rev-range-incoming.
> ERROR: --end-revision cannot be combined with --no-range!