git-{sinceby,{by,log}userrange}-command: Reimplementation: Explicitly parse known log arguments instead of parsing the git-*-command GIT-COMMAND args (cp. a4fbf68b)
As that is very cumbersome, and doesn't handle nested git-*-command (which would require recursion). Additionally (and the trigger for the rewrite), in the "git sinceby ds" command, any DIFF-ARGUMENT(s) actually should only go to the "ds" command, but not the log command that creates the range.
Drop support for git-sinceby-command --pass-file-args; the implementation is wrong (files are passed twice with --with-log-args), and it's currently not used at all. And for consistency, the other commands should support that as well.

(This relates to commit a4fbf68b2ad844d9e7fd67883dee941b642f57e0)