git-{{byuser,log}range,segregated-user,sinceby}-command: CHG: Split --with-log-args-except-last, add --files-for-{range,command}
Only git-sinceby-command parsed off any FILE(s), and then allowed to pass them to the GIT-COMMAND via --pass-file-args (which wasn't used anywhere).
It's better to pass FILE(s) along (like the other commands which are completely ignorant of FILE args) by default, and offer (because it's trivial) flags to shut off FILE argument passing to either the Git command that determines the range, or the final GIT-COMMAND.
DWIM: With this, for example commit message filtering on a subdir then also logs just the subdir's changes.
The importance of a special last argument (e.g. for my *by and *changelog commands) had already been diminished by explicitly parsing git log arguments instead of taking everything.
With the file parsing, this now is a fully separate concern from whether log arguments are passed to the range command or not. Therefore, introduce a separate --except-last argument and drop --with-log-args-except-last.