git-others-command: BUG: Individual author args need to be escaped, too; add --literal flag
A --author="dependabot[bot]" doesn't match, as it's interpreted as a pattern. We need regexp escaping not just when joining into a single pattern.
However, as there are options like --name that may be used for literal consumption, make this configuable via a new --literal flag. It's listed here as orthogonal to --joined-as-pattern (as only that makes sense), but actually would work in conjunction as well.