git-files-command: FIX: Duplicate --, need to parse FILE(s) off only after -N or +N have captured their arguments
Otherwise it's impossible to pass a -- separator to the GIT-COMMAND via -N or +N. It would be recognized as the file separator and following (non-FILE!) arguments be taken as files.
Example:
$ git-rev-range ... files-command --source-exec showfiles RANGE \; +2 -- FILES diffselected --log-range RANGE

For plain appending, a -- is prepended automatically to the files; with the parsed-off --, that would make two, which can cause problems with (the messy) commandOnSelected handling. (I've noticed then during debugging.)
For --keep-position, no -- is used, nor can it be passed. This alternative is meant for custom separation (e.g. between \; .... \;)