Generalization: Renaming: Extract git-dtag
Cannot keep "dt" because that clashes with my difftool shortening. This isn't a common scope, so slightly lengthening it to "dtag" should be okay, and it better signifies the subject (tags).
Introducing new scopeArgSyphon customization to move the special git-tagdelta-command arguments to the correct position.
The usage help also is special (it's either-or with the -r|--revision REVISION), so introduce a scopeArgsOverride.