Handle aggregate emaill{og,c}
We want a single email across all refs, so email-command has to be the topmost command.
For the logging, we delegate back to the br-variant.
git-email-command needs a little tweak to actually take the passed GIT-COMMAND instead of defaulting to "log".
For emaillc, the diff output needs to be turned off.