FIX: git-[lou]bm actually support inout
But as both branch and master are passed to inout, and git-rev-range is part of scopeCommand and therefore cannot be removed, we need some magic to have the two refs as separate arguments and in the right order.
git-rev-range --with-range ' ' now renders END-REVISION and REVISION as two separate arguments.
git-rev-range --reverse swaps both (so that we compare in the usual BRANCH <-> master order).