CHG: Rename stash-i to stashst and only launch editor when "push" is given to stash{i,st,uu,hunk}
stash-i is a misnomer because it stashes both index and unstaged changes, and just keeps the index intact. As written, this is great for testing without working copy (to verify that the staged tests are now failing), but not a "workingcopy-only" stage.
Rename to stashst(atus); it's not 100% correct but like my "git st" indicates a combination of index and unstaged.
Also rename the {correctto,stashrelate}-i aliases to *st. Dunno how useful they are; I've added them mostly for completeness.