ENH: Add git-stashI for committing just the working copy
"git stashst" also stashes the index, but keeps it; it's meant for accumulative committing, but I had working-copy only stashing in mind when I wrote it (as git-stashI).
I can actually implement working copy-only stashing easily (but a bit convoluted, and with the risk of patch application errors) by leveraging my swapstage command and then delegating to stashi, and finally putting the (now unstaged) changes back into the index.