FIX: Need special implementation for lgfilesby to only offer authors from the range
The problem here is that the authors are needed twice: To limit the file selection, and to log only commits from them. With the GIT_SELECTED_COMMAND_DEFAULT_FILES, that cannot be composed.
Instead, resolve the --author arguments first; a new git-quoted variant of git-echo allows us to collect them in quoted form, plus the (resolved) revision range as well.
Then install that filter into the default files for the selected-command, and eventually execute lg with that filter, too.