l{[o]g,c}lastfiles{mine,team,by}: FIX: Only worked when the last commit is by that author
We cannot simply tack a --author=... to loglastfiles; it'll still just target the very last commit, regardless of its author. Instead, we have to find the last commit by the desired author, and use that as the start revision for loglastfiles.
"byuserrange-command --last-only" gives us the last commit of the passed author.
We pass that to showfiles-command via --revision RANGE to get the correct files.
The RANGE also needs to be passed to the eventual log command (showfiles-command just appends the files), so that it starts from the right commit.