git-logrange-command:CHG: Make --revision REVISION the start revision for a search for the first matching commit (cp. 4f5bcf95)
Instead of complaining that the first revision already doesn't match. That isn't so useful. I want to point the search to the beginning of e.g. a merged branch and have it find the consecutive matching commits there.

(This relates to commit 4f5bcf95a5cbeabccb20fe7a49242bbb757d6b41)