branch-range: FIX: Actually use the merge base for pred* and the last commit on the branch for base*
Simply using the $scopeRevision (e.g. "master" for git-bm) is wrong when that branch has progressed. We need the first commit made to the merge base. Let's use rnlog --max-count 1 to obtain that; it's cumbersome to go against the history from the merge-base.
For pred*, --one-more likewise is ineffective when branches have diverged; here, we need to obtain the merge-base and then use that commit reference.