git-unmerge: Robustness: Check for changes to the merge commit itself and ask the user whether to discard those
Changes might be directly amended to a merge commit; these would be lost by unmerging the branch. Detect (any) changes and alert the user. It could also just be conflict resolution changes (which would be safe to discard and nothing more than a hassle to reconcile again later (where rerere might help)).
Though it seems possible to detect the difference (looking at the preimage for conflict markers <<<<<<< and >>>>>>> and nothing else), this is deemed not worthwhile and too brittle. We just dump the diff to the user and let them decide.