Just filter by name and ignore email if --name is given
I've used a post-filter (i.e. only extract the desired part after the selection) so that the email lookup could be done by name (as the email address might be cryptic or use a (funny) alias) as well.
But when just the name is desired, it doesn't make sense to show multiple entries with the same name but differing email addresses (which happens frequently, e.g. for users that have personal and GitHub @users.noreply.github.com addresses).

For the commandOnSelected implementation, this is a simple change from --post-filter to --filter.
For the fallback implementation, define both candidateFilter and printFilter, and place personFilter accordingly.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: git version 2.47.0, hub version 2.14.2, gh version 2.55.0 (2024-08-20), git-lfs/3.5.1 (GitHub; linux amd64; go 1.21.8)