git-others-command: Robustness: Also match shorter and longer authors, and consider the email part
If another author is a prefix of the excluded author, it so far was excluded as well, because every character position doesn't match. We need to accept shorter authors unconditionally.
Likewise, if another author has the excluded author as its prefix, every character position doesn't match, so larger authors need to be matched as well.
For these matches, it's important that we consider the optional email part. (Git's --authors always matches against the full NAME <EMAIL>; in git-others-commands, the $me is the full author, but when invoked by git-authors-command --invert-authors, only the name is used.)
Check whether the author contains the email and if not assert on the / <.+>$/

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: git version 2.47.0, hub version 2.14.2, gh version 2.55.0 (2024-08-20), git-lfs/3.6.0 (GitHub; linux amd64; go 1.23.1)