git-commit: ENH: Allow recording of commit to be ignored by git blame via --skip-blame
It makes sense to commit the .git-blame-ignore-revs file; unfortunately, this cannot be done in the to-be-ignored commit itself, as the hash is part of it. So there needs to be a separate housekeeping commit. These commits also shouldn't be rebased or amended.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: git version 2.47.1, hub version 2.14.2, gh version 2.55.0 (2024-08-20), git-lfs/3.6.0 (GitHub; linux amd64; go 1.23.1)