Refactoring: Consistency: Replace the derivation of the aggregateCommand with explicit passing
With git- vs. hub-, this is too complex for metaprogramming, and not worthwhile.
In overtimeCommandExtension(), I've been forced to pass it in, because of the need to support {mine,...} aliases; better do this consistently.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: git version 2.47.1, hub version 2.14.2, gh version 2.55.0 (2024-08-20), git-lfs/3.6.0 (GitHub; linux amd64; go 1.23.1)