ENH: Add --external-git-dir syntactic sugar around --separate-git-dir
I need this to create (temporary) Git repositories for working on agorum core customizations on the mounted root Samba share.
The whole metadata must be outside that (slow) mount, on the local system. The location should be auto-generated from the working copy name, so that it's quick to setup.
If the external gitdir already exists, the implementation doesn't check that, and "git init" just re-initializes that repo. That might be useful, or maybe I want a check and choose an incremented name instead. We'll see; because of the message it should be fairly noticeable when it occurs.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: git version 2.49.0, hub version 2.14.2, gh version 2.70.0 (2025-04-15), git-lfs/3.6.1 (GitHub; linux amd64; go 1.23.3)